requirements.txt
torch
torchvision
numpy
pillow
opencv-python-headless
scikit-learn
fastapi
uvicorn
streamlit
requests
